refactor: divide the schema into modules
